+++ src/main/kotlin/id/walt/services/vc/WaltIdJsonLdCredentialService.kt
@@ -13,11 +13,7 @@ import id.walt.services.essif.TrustedIssuerClient.domain
 import id.walt.services.keystore.KeyStoreService
 import id.walt.signatory.ProofConfig
 import id.walt.signatory.ProofType
-import id.walt.vclib.Helpers.encode
-import id.walt.vclib.Helpers.toCredential
-import id.walt.vclib.VcLibManager
-import id.walt.vclib.VcUtils
-
+import id.walt.vclib.model.toCredential
 import id.walt.vclib.credentials.VerifiableAttestation
 import id.walt.vclib.credentials.VerifiablePresentation
 import id.walt.vclib.model.CredentialSchema
@@ -193,7 +189,7 @@ open class WaltIdJsonLdCredentialService : JsonLdCredentialService() {
     fun verifyVerifiablePresentation(json: String) =
         VerificationResult(verifyVp(json), VerificationType.VERIFIABLE_PRESENTATION)
 
-    override fun verify(vcOrVp: String): VerificationResult = when (VcLibManager.getVerifiableCredential(vcOrVp)) {
+    override fun verify(vcOrVp: String): VerificationResult = when (VerifiableCredential.fromString(vcOrVp)) {
         is VerifiablePresentation -> verifyVerifiablePresentation(vcOrVp)
         else -> verifyVerifiableCredential(vcOrVp)
     }
@@ -204,7 +200,7 @@ open class WaltIdJsonLdCredentialService : JsonLdCredentialService() {
 
         val vcObj = vcJson.toCredential()
 
-        val issuer = VcUtils.getIssuer(vcObj)
+        val issuer = vcObj.issuer!!
         log.debug { "VC decoded: $vcObj" }
 
         val vcVerified = verifyVc(issuer, vcJson)
@@ -319,7 +315,7 @@ open class WaltIdJsonLdCredentialService : JsonLdCredentialService() {
             issuer = "did:ebsi:2757945549477fc571663bee12042873fe555b674bd294a3",
             issuanceDate = "2019-06-22T14:11:44Z",
             validFrom = "2019-06-22T14:11:44Z",
-            credentialSubject = VerifiableAttestation.CredentialSubject(
+            credentialSubject = VerifiableAttestation.VerifiableAttestationSubject(
                 id = "id123"
             ),
             credentialStatus = CredentialStatus(
@@ -378,7 +374,7 @@ open class WaltIdJsonLdCredentialService : JsonLdCredentialService() {
 
             if (it is VerifiablePresentation) return true
 
-            val credentialSchemaUrl = VcUtils.getCredentialSchemaUrl(it)
+            val credentialSchemaUrl = it.credentialSchema
 
             if (credentialSchemaUrl == null) {
                 log.debug { "Credential has no associated credentialSchema property" }