+++ src/main/kotlin/id/walt/signatory/dataproviders/DefaultDataProvider.kt
@@ -1,0 +1,11 @@
+package id.walt.signatory.dataproviders
+
+import id.walt.signatory.ProofConfig
+import id.walt.vclib.model.VerifiableCredential
+
+object DefaultDataProvider : AbstractDataProvider<VerifiableCredential>() {
+    override fun populateCustomData(vc: VerifiableCredential, proofConfig: ProofConfig): VerifiableCredential {
+        // nothing to do
+        return vc
+    }
+}