+++ src/main/kotlin/id/walt/signatory/rest/SignatoryController.kt
@@ -1,9 +1,9 @@
 package id.walt.signatory.rest
 
-import id.walt.signatory.MergingDataProvider
+import id.walt.signatory.dataproviders.MergingDataProvider
 import id.walt.signatory.ProofConfig
 import id.walt.signatory.Signatory
-import id.walt.vclib.Helpers.encode
+
 import io.javalin.http.Context
 import io.javalin.plugin.openapi.dsl.document