+++ src/test/kotlin/id/walt/rest/EssifApiTest.kt
@@ -59,7 +59,7 @@ class EssifApiTest : AnnotationSpec() {
         resp shouldStartWith "0x"
     }
 
-    @Test
+    //@Test // disabled for now, as it times out on github builds
     fun testTimestampByTxHash() = runBlocking {
         val resp = client.get<String>("$ESSIF_API_URL/v1/client/timestamp/txhash/0x42348e1ee94cc78d5e5494f71b502416aa566b626151f8dee333804f061bda1d") {
             contentType(ContentType.Application.Json)