+++ src/test/kotlin/id/walt/signatory/SignatoryApiTest.kt
@@ -7,8 +7,8 @@ import id.walt.services.did.DidService
 import id.walt.signatory.rest.IssueCredentialRequest
 import id.walt.signatory.rest.SignatoryRestAPI
 import id.walt.test.RESOURCES_PATH
-import id.walt.vclib.Helpers.encode
-import id.walt.vclib.Helpers.toCredential
+
+import id.walt.vclib.model.toCredential
 import id.walt.vclib.credentials.*
 import id.walt.vclib.templates.VcTemplateManager
 import io.github.rybalkinsd.kohttp.dsl.httpPost
@@ -254,7 +254,7 @@ class SignatoryApiTest : AnnotationSpec() {
         println(vc)
         val cred = vc?.toCredential() as VerifiableAuthorization
         cred.issuer shouldBe did
-        cred.credentialSubject.id shouldBe did
+        cred.credentialSubject?.id shouldBe did
         cred.proof?.verificationMethod shouldBe "$did#key-1"
     }