+++ src/test/kotlin/id/walt/test/DummySignatoryDataProvider.kt
@@ -1,6 +1,7 @@
 package id.walt.test
 
 import id.walt.signatory.*
+import id.walt.signatory.dataproviders.DefaultDataProvider
 import id.walt.vclib.credentials.VerifiableDiploma
 import id.walt.vclib.credentials.VerifiableId
 import id.walt.vclib.model.Proof
@@ -9,7 +10,7 @@ import id.walt.vclib.model.VerifiableCredential
 class DummySignatoryDataProvider : SignatoryDataProvider {
     override fun populate(template: VerifiableCredential, proofConfig: ProofConfig): VerifiableCredential {
         if (template is VerifiableId) {
-            VerifiableIdDataProvider().populate(template, proofConfig)
+            DefaultDataProvider.populate(template, proofConfig)
             template.evidence!!.id = "Dummy test value (waiting for EBSI schema update)"
             if (proofConfig.proofType == ProofType.JWT) template.proof = Proof(
                 type = "Ed25519Signature2018", // Dummy test value (waiting for EBSI schema update)
@@ -23,7 +24,7 @@ class DummySignatoryDataProvider : SignatoryDataProvider {
             )
         }
         if (template is VerifiableDiploma) {
-            VerifiableDiplomaDataProvider().populate(template, proofConfig)
+            DefaultDataProvider.populate(template, proofConfig)
             template.credentialSubject!!.awardingOpportunity!!.awardingBody.eidasLegalIdentifier = "Dummy test value (waiting for EBSI schema update)"
             if (proofConfig.proofType == ProofType.JWT) template.proof = Proof(
                 type = "Ed25519Signature2018", // Dummy test value (waiting for EBSI schema update)